# packages/backend/src/server.ts
import express from 'express';
import swaggerUi from 'swagger-ui-express';
import { swaggerSpec } from './swagger';
import {
  initDb,
  saveCommits,
  saveSummary,
  getSummary,
  getCommits,
  getLatestSummaryBeforeDate,
  StoredCommit,
  BulkCommitPayloadSchema,
} from './lib/db';
import { generateDailySummary } from './lib/llm';

import rateLimit from 'express-rate-limit';

const app = express();
const PORT = process.env.PORT ? parseInt(process.env.PORT, 10) : 4001;

// Trust reverse proxy (Railway, Render, Nginx) for accurate IP resolution
app.set('trust proxy', 1);

app.use(express.json());

// Rate Limiter: 10 requests per hour per IP (temporary measure to prevent spam)
const apiLimiter = rateLimit({
  windowMs: 60 * 60 * 1000, // 1 hour
  limit: 10, // Limit each IP to 10 requests per windowMs
  standardHeaders: 'draft-7', // Return standard RateLimit headers
  legacyHeaders: false,
  message: {
    error: 'Too Many Requests',
    message: 'Rate limit exceeded: Maximum 10 requests per hour. Please try again later.',
  },
});

// Apply rate limiter to API routes
app.use('/api', apiLimiter);

// Swagger UI & JSON Spec (exempt from rate limit)
app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerSpec));
app.get('/api-docs.json', (_req, res) => {
  res.setHeader('Content-Type', 'application/json');
  res.send(swaggerSpec);
});

initDb();

app.post('/api/commits', async (req, res) => {
  const parsed = BulkCommitPayloadSchema.safeParse(req.body);

  if (!parsed.success) {
    return res.status(400).json({
      error: 'Invalid payload',
      details: parsed.error.format(),
    });
  }

  const payload = parsed.data;

  try {
    await saveCommits(payload);
    const dayCommits = await getCommits(payload.userId, payload.date);

    const summaryText = await generateDailySummary(
      payload.userId,
      payload.date,
      dayCommits,
      payload.template,
      payload.otherActivities
    );

    const totalCommits = dayCommits.length;

    console.log('Summary generated successfully!', summaryText);

    await saveSummary(payload.userId, payload.date, summaryText, totalCommits);

    const byRepo = new Map<string, StoredCommit[]>();
    for (const entry of dayCommits) {
      const list = byRepo.get(entry.repoName) || [];
      list.push(entry);
      byRepo.set(entry.repoName, list);
    }

    const repos: {
      name: string;
      commits: { hash: string; message: string }[];
    }[] = [];

    for (const [repoName, entries] of byRepo.entries()) {
      const seen = new Set<string>();
      const uniqueCommits: { hash: string; message: string }[] = [];
      for (const c of entries) {
        if (!seen.has(c.hash)) {
          seen.add(c.hash);
          uniqueCommits.push({
            hash: c.hash,
            message: c.message,
          });
        }
      }

      repos.push({
        name: repoName,
        commits: uniqueCommits,
      });
    }

    // Fetch yesterday's summary relative to the new date
    const lastSummary = await getLatestSummaryBeforeDate(payload.userId, payload.date);
    let yesterday = null;
    if (lastSummary) {
      yesterday = {
        date: lastSummary.date,
        summary: lastSummary.summary,
        totalCommits: lastSummary.totalCommits,
      };
    }

    return res.json({
      userId: payload.userId,
      date: payload.date,
      summary: summaryText,
      repos,
      yesterday, // <-- New addition
    });
  } catch (error) {
    console.error('Error processing commits:', error);
    return res.status(500).json({ error: 'Internal Server Error' });
  }
});

app.get('/api/daily-summary', async (req, res) => {
  const userId = (req.query.userId as string) || '';
  const date = (req.query.date as string) || '';
  if (!userId || !date) {
    return res.status(400).json({
      error: 'Missing userId or date',
    });
  }

  try {
    const storedSummary = await getSummary(userId, date);
    const dayCommits = await getCommits(userId, date);

    const byRepo = new Map<string, StoredCommit[]>();
    for (const entry of dayCommits) {
      const list = byRepo.get(entry.repoName) || [];
      list.push(entry);
      byRepo.set(entry.repoName, list);
    }

    const repos: {
      name: string;
      commits: { hash: string; message: string }[];
    }[] = [];

    for (const [repoName, entries] of byRepo.entries()) {
      const seen = new Set<string>();
      const uniqueCommits: { hash: string; message: string }[] = [];
      for (const c of entries) {
        if (!seen.has(c.hash)) {
          seen.add(c.hash);
          uniqueCommits.push({
            hash: c.hash,
            message: c.message,
          });
        }
      }

      repos.push({
        name: repoName,
        commits: uniqueCommits,
      });
    }

    return res.json({
      userId,
      date,
      summary: storedSummary?.summary || 'No summary generated yet.',
      repos,
    });
  } catch (error) {
    console.error('Error fetching summary:', error);
    return res.status(500).json({ error: 'Internal Server Error' });
  }
});

app.get('/api/recent-summaries', async (req, res) => {
  console.log('Received request for recent summaries');
  const userId = (req.query.userId as string) || '';
  const dateStr = req.query.date as string;
  console.log(dateStr);
  if (!userId) {
    return res.status(400).json({ error: 'Missing userId' });
  }

  try {
    const [todaySummary, lastSummary] = await Promise.all([
      getSummary(userId, dateStr),
      getLatestSummaryBeforeDate(userId, dateStr),
    ]);

    let yesterdayDate = null;
    if (lastSummary) {
      yesterdayDate = lastSummary.date;
    }

    return res.json({
      userId,
      today: {
        date: dateStr,
        summary: todaySummary?.summary || null,
        totalCommits: todaySummary?.totalCommits || 0,
      },
      yesterday: {
        date: yesterdayDate || 'N/A',
        summary: lastSummary?.summary || null,
        totalCommits: lastSummary?.totalCommits || 0,
      },
    });
  } catch (error) {
    console.error('Error fetching recent summaries:', error);
    return res.status(500).json({ error: 'Internal Server Error' });
  }
});

// Healthcheck
app.get('/health', (_req, res) => {
  res.json({ status: 'ok' });
});

app.listen(PORT, '0.0.0.0', () => {
  console.log(`LogMyCode backend listening on port ${PORT}`);
  console.log(`Swagger documentation available at http://localhost:${PORT}/api-docs`);
});